refactor(chat): Complete chat consolidation into unified component
Merged three separate chat implementations (exercise, lesson, and
course chat) into a single, feature-rich ChatInterface component
with configurable features via props.

Key changes:
- Unified ChatInterface now supports all features: media upload,
  quick actions (hint/solution), reset button, math tools, and
  mobile toggle
- Added lazy loading for math components (FormulaPanel,
  MathPalette) to avoid loading them when not needed
- Updated all three usage locations to use the unified component
- Removed duplicate code (~500+ lines eliminated)
- Fixed syntax error in onClick handler (})  vs }})

Files changed:
- src/ui/web/chat/ChatInterface/index.tsx: Enhanced with all features
- src/ui/web/chat/index.ts: Export ViewMode type
- lesson page and exercise page: Updated to use unified component
- Deleted old exercise ChatInterface, NotebookChat wrapper, and test

Benefits:
- Single source of truth for chat functionality
- Easier maintenance and feature additions
- Consistent UX across all chat contexts
- Reduced bundle size through lazy loading
