--- src/app/(frontend)/courses/[courseSlug]/chapters/[chapterSlug]/lessons/[lessonSlug]/exercises/[exerciseId]/_components/ChatInterface/index.tsx
@@ -1,365 +1,0 @@
-'use client'
-
-import { ChatMessageRole } from '@/infra/llm/chat-message-role'
-import { useTranslations } from '@/ui/web/providers/I18n'
-import { cn } from '@/infra/utils/ui'
-import {
-  BookOpen,
-  Loader2,
-  Plus,
-  Send,
-  MessageSquare,
-  FileText,
-  X,
-  Image as ImageIcon,
-  FileUp,
-} from 'lucide-react'
-import React, { useEffect, useState } from 'react'
-import { ChatMessageContent, useNotebookChat } from '@/ui/web/chat'
-import { FormulaPanel } from '../FormulaPanel'
-import { MathPalette } from '../MathPalette'
-import type { ViewMode } from '../ExerciseWorkspace/exercise-workspace-types'
-
-interface ChatInterfaceProps {
-  exerciseId?: string
-  lessonId?: string
-  // NEW: Mobile mode support
-  displayMode?: 'full' | 'input-only'
-  onChatInteraction?: () => void
-  // Mobile toggle
-  isMobile?: boolean
-  viewMode?: ViewMode
-  onModeToggle?: () => void
-}
-
-export function ChatInterface({
-  exerciseId,
-  lessonId,
-  displayMode = 'full',
-  onChatInteraction,
-  isMobile,
-  viewMode,
-  onModeToggle,
-}: ChatInterfaceProps) {
-  const t = useTranslations('courses')
-  const {
-    messages,
-    inputValue,
-    isLoading,
-    isLoadingHistory,
-    messagesContainerRef,
-    messagesEndRef,
-    inputRef,
-    fileInputRef,
-    setInputValue,
-    handleSubmit,
-    // Media upload
-    uploadedMedia,
-    isUploading,
-    handleFileSelect,
-    removeMedia,
-    openFilePicker,
-  } = useNotebookChat({
-    initialMessage: t('chatWelcome'),
-    authRequiredMessage: t('chatAuthRequired'),
-    errorMessage: t('chatError'),
-    hintPrompt: t('chatHintPrompt'),
-    solutionPrompt: t('chatSolutionPrompt'),
-    fullSolutionPrompt: t('chatFullSolutionPrompt'),
-    resetConfirmMessage: t('chatResetConfirm'),
-    resetSuccessMessage: t('chatResetSuccess'),
-    resetErrorMessage: t('chatResetError'),
-    acknowledgment: t('chatAIAcknowledgment'),
-    exerciseId,
-    lessonId,
-    // Media upload messages
-    unsupportedFileTypeMessage: t('chatUnsupportedFileType'),
-    fileTooLargeMessage: t('chatFileTooLarge'),
-    maxFilesMessage: t('chatMaxFiles'),
-    uploadFailedMessage: t('chatUploadFailed'),
-  })
-
-  const [isMathPaletteOpen, setIsMathPaletteOpen] = useState(false)
-  const [isFormulaPanelOpen, setIsFormulaPanelOpen] = useState(false)
-  const [mathPreview, setMathPreview] = useState('')
-
-  // Update LaTeX preview
-  useEffect(() => {
-    if (inputValue.includes('\\') || inputValue.includes('^')) {
-      setMathPreview(inputValue)
-    } else {
-      setMathPreview('')
-    }
-  }, [inputValue])
-
-  const injectFormula = (template: string, cursorOffset: number) => {
-    if (!inputRef.current) return
-
-    const start = inputRef.current.selectionStart ?? 0
-    const end = inputRef.current.selectionEnd ?? 0
-    const before = inputValue.substring(0, start)
-    const after = inputValue.substring(end)
-
-    const newValue = before + template + after
-    setInputValue(newValue)
-
-    // Move cursor after state update
-    requestAnimationFrame(() => {
-      const newCursorPos = start + cursorOffset
-      inputRef.current?.setSelectionRange(newCursorPos, newCursorPos)
-      inputRef.current?.focus()
-    })
-  }
-
-  // Handle input change (no expansion while typing)
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setInputValue(e.target.value)
-  }
-
-  const handleFormSubmit = (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsMathPaletteOpen(false)
-    setIsFormulaPanelOpen(false)
-
-    // Trigger interaction callback on submit
-    if (onChatInteraction) {
-      onChatInteraction()
-    }
-
-    handleSubmit(e)
-  }
-
-  return (
-    <div className="flex flex-col h-full min-h-0">
-      {/* Messages Area - Hidden when displayMode is 'input-only' */}
-      <div
-        ref={messagesContainerRef}
-        className={cn(
-          'flex-grow overflow-y-auto p-5 space-y-4 min-h-0',
-          displayMode === 'input-only' && 'hidden',
-        )}
-      >
-        {isLoadingHistory && (
-          <div className="flex items-center justify-center p-4 text-muted-foreground text-sm">
-            <Loader2 className="w-4 h-4 animate-spin mr-2" />
-            {t('chatLoadingHistory')}
-          </div>
-        )}
-        {!isLoadingHistory &&
-          messages.map((msg, idx) => {
-            console.log('[ChatInterface] Message:', {
-              idx,
-              role: msg.role,
-              hasMedia: !!msg.media,
-              mediaLength: msg.media?.length,
-              media: msg.media,
-            })
-            return (
-              <div
-                key={idx}
-                className={cn(
-                  'max-w-[85%] px-[18px] py-3.5 text-base leading-relaxed shadow-sm',
-                  msg.role === ChatMessageRole.User
-                    ? 'ml-auto bg-primary text-primary-foreground rounded-[20px] rounded-bl-[4px]'
-                    : 'mr-auto bg-card text-foreground border border-border rounded-[20px] rounded-br-[4px]',
-                )}
-              >
-                {msg.media && msg.media.length > 0 && (
-                  <div className="flex flex-wrap gap-1.5 mb-2">
-                    {msg.media.map((mediaItem, mediaIdx) => (
-                      <div
-                        key={mediaIdx}
-                        className={cn(
-                          'inline-flex items-center gap-1.5 px-2 py-1 rounded-full text-xs',
-                          msg.role === ChatMessageRole.User
-                            ? 'bg-primary-foreground/20'
-                            : 'bg-muted',
-                        )}
-                      >
-                        <ImageIcon className="w-3 h-3" />
-                        <span className="max-w-[120px] truncate">
-                          {mediaItem.filename || `media-${mediaIdx + 1}`}
-                        </span>
-                      </div>
-                    ))}
-                  </div>
-                )}
-                <ChatMessageContent content={msg.content} />
-              </div>
-            )
-          })}
-        {isLoading && (
-          <div className="mr-auto bg-card text-foreground border border-border px-[18px] py-3.5 rounded-[20px] rounded-br-[4px] max-w-[85%] flex items-center gap-2 shadow-sm">
-            <Loader2 className="w-4 h-4 animate-spin" />
-            <span>{t('chatThinking')}</span>
-          </div>
-        )}
-        <div ref={messagesEndRef} />
-      </div>
-
-      {/* Input Container - Always rendered */}
-      <div className="flex-grow-0 flex-shrink-0 bg-card border-t border-border p-5 pb-8 relative">
-        {/* Math Preview Popup */}
-        {mathPreview && (
-          <div className="absolute bottom-full left-5 right-5 mb-2.5 bg-card border border-primary-soft rounded-xl p-2.5 text-center shadow-panel z-20">
-            <span className="text-sm font-mono">{mathPreview}</span>
-          </div>
-        )}
-
-        {/* Formula Panel (Popup) */}
-        <FormulaPanel
-          isOpen={isFormulaPanelOpen}
-          onClose={() => setIsFormulaPanelOpen(false)}
-          onInject={injectFormula}
-        />
-
-        {/* Math Palette (Slide-out) */}
-        <MathPalette isOpen={isMathPaletteOpen} onInject={injectFormula} />
-
-        {/* Toolbar Above Input */}
-        <div className="flex gap-4 mb-2.5 px-1.5 justify-between items-center">
-          <button
-            type="button"
-            className={cn(
-              'p-1 text-muted-foreground hover:text-primary transition-colors',
-              isFormulaPanelOpen && 'text-primary',
-            )}
-            onClick={() => {
-              setIsFormulaPanelOpen(!isFormulaPanelOpen)
-              setIsMathPaletteOpen(false)
-            }}
-            aria-label={t('formulaSheet')}
-          >
-            <BookOpen className="w-5 h-5" />
-          </button>
-
-          {/* Mobile Toggle - Shows opposite mode */}
-          {isMobile && viewMode && onModeToggle && (
-            <button
-              type="button"
-              className="flex items-center gap-1.5 px-3 py-1.5 rounded-md text-sm font-medium transition-colors bg-muted hover:bg-muted/80"
-              onClick={onModeToggle}
-              aria-label={viewMode === 'PDF' ? t('switchToChat') : t('switchToPdf')}
-            >
-              {viewMode === 'PDF' ? (
-                <>
-                  <MessageSquare className="w-4 h-4" />
-                  <span className="hidden sm:inline">{t('chat')}</span>
-                </>
-              ) : (
-                <>
-                  <FileText className="w-4 h-4" />
-                  <span className="hidden sm:inline">{t('content')}</span>
-                </>
-              )}
-            </button>
-          )}
-        </div>
-
-        {/* Media Preview Chips */}
-        {uploadedMedia.length > 0 && (
-          <div className="flex flex-wrap gap-2 mb-2.5 max-w-[850px] mx-auto">
-            {uploadedMedia.map((media) => (
-              <div
-                key={media.id}
-                className="flex items-center gap-1.5 bg-muted rounded-full px-3 py-1.5 text-sm border border-input"
-              >
-                {media.mimeType.startsWith('image/') ? (
-                  <ImageIcon className="w-4 h-4 text-muted-foreground" />
-                ) : (
-                  <FileUp className="w-4 h-4 text-muted-foreground" />
-                )}
-                <span className="max-w-[120px] truncate text-foreground">{media.filename}</span>
-                <button
-                  type="button"
-                  onClick={() => removeMedia(media.id)}
-                  className="p-0.5 hover:bg-destructive/20 rounded-full transition-colors"
-                  aria-label={t('chatRemoveFile')}
-                >
-                  <X className="w-3.5 h-3.5 text-muted-foreground hover:text-destructive" />
-                </button>
-              </div>
-            ))}
-          </div>
-        )}
-
-        {/* Input Wrapper */}
-        <form onSubmit={handleFormSubmit}>
-          <div className="max-w-[850px] mx-auto bg-muted rounded-[30px] flex items-center px-4 py-1.5 border border-input gap-3">
-            <input
-              ref={inputRef}
-              type="text"
-              className="flex-1 bg-transparent border-none outline-none py-2.5 text-[17px] text-foreground placeholder:text-muted-foreground"
-              placeholder={t('chatInputPlaceholder')}
-              value={inputValue}
-              onChange={handleInputChange}
-              onKeyDown={(e) => {
-                // Let the form onSubmit handle Enter key submission
-                // Don't call handleKeyDown to avoid double submission
-                if (e.key === 'Enter' && !e.shiftKey) {
-                  e.preventDefault()
-                  handleFormSubmit(e as unknown as React.FormEvent)
-                }
-              }}
-              disabled={isLoading}
-            />
-
-            {/* Math Keyboard Toggle */}
-            <button
-              type="button"
-              className={cn(
-                'p-1.5 text-muted-foreground hover:text-primary transition-colors',
-                isMathPaletteOpen && 'text-primary',
-              )}
-              onClick={() => {
-                setIsMathPaletteOpen(!isMathPaletteOpen)
-                setIsFormulaPanelOpen(false)
-              }}
-              aria-label={t('mathKeyboard')}
-            >
-              <span className="text-lg font-bold">ƒ</span>
-            </button>
-
-            {/* File Upload */}
-            <button
-              type="button"
-              className={cn(
-                'p-1.5 text-muted-foreground hover:text-primary transition-colors',
-                isUploading && 'opacity-50 cursor-not-allowed',
-              )}
-              onClick={openFilePicker}
-              disabled={isUploading || uploadedMedia.length >= 5}
-              aria-label={t('chatAttachFile')}
-            >
-              {isUploading ? (
-                <Loader2 className="w-5 h-5 animate-spin" />
-              ) : (
-                <Plus className="w-5 h-5" />
-              )}
-            </button>
-            <input
-              ref={fileInputRef}
-              type="file"
-              className="hidden"
-              accept="image/jpeg,image/jpg,image/png,image/webp,application/pdf"
-              multiple
-              onChange={(e) => handleFileSelect(e.target.files)}
-            />
-
-            {/* Send Button */}
-            <button
-              type="submit"
-              className="w-10 h-10 rounded-full bg-primary text-primary-foreground flex items-center justify-center shadow-input hover:bg-primary/90 transition-all hover:scale-105 disabled:opacity-50 disabled:cursor-not-allowed"
-              disabled={
-                isLoading || isUploading || (!inputValue.trim() && uploadedMedia.length === 0)
-              }
-              aria-label={t('sendMessage')}
-            >
-              <Send className="w-5 h-5" />
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  )
-}
