+++ src/ui/web/chat/index.ts
@@ -1,3 +1,3 @@
 export { ChatMessageContent } from './ChatMessageContent'
-export { ChatInterface } from './ChatInterface'
+export { ChatInterface, type ViewMode } from './ChatInterface'
 export { useNotebookChat } from './hooks/useNotebookChat'
